Rename net/fetch/request.rs to net/fetch/methods.rs and move Request to net_traits
