+++ components/net/lib.rs
@@ -46,6 +46,6 @@ pub mod websocket_loader;
 pub mod fetch {
     #![allow(dead_code, unused)] // XXXManishearth this is only temporary until the Fetch mod starts being used
     pub mod cors_cache;
-    pub mod request;
+    pub mod methods;
     pub mod response;
 }
