+++ components/net_traits/lib.rs
@@ -38,6 +38,7 @@ use websocket::header;
 pub mod hosts;
 pub mod image_cache_thread;
 pub mod net_error_list;
+pub mod request;
 pub mod response;
 pub mod storage_thread;
